gym-maze/requirements.txt
numpy~=1.24.2
gym~=0.26.2
requests~=2.28.2
pygame~=2.2.0
setuptools~=65.5.1
PyJWT~=2.6.0
jwcrypto~=1.4.2
cryptography~=39.0.2